requirements.txt
Flask==2.3.4
openai==1.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
